Add configurable audio quality and send toggle (brief §3.3)
Slave can now choose sample rate (48/24kHz) and bit depth (16/8-bit)
and toggle whether to send audio at all, from a new Audio tab on the
Discovery screen. Master adapts playback to whatever format the Slave
announces. Required switching the audio pipeline (jitter buffer,
protocol framing, channel, capture/playback) from fixed int16_t
samples to a runtime AudioFormat over raw bytes, and fixing an 8-bit
silence-fill bug found during that refactor (0x80 midpoint, not 0x00).

+++ src/audio/jitter_buffer.cpp
@@ -15,13 +15,13 @@ constexpr size_t kMaxConsecutiveMisses = 40;
 
 JitterBuffer::JitterBuffer(size_t target_depth) : target_depth_(target_depth) {}
 
-void JitterBuffer::push(uint32_t seq, std::vector<int16_t> frame) {
+void JitterBuffer::push(uint32_t seq, std::vector<uint8_t> frame) {
   std::lock_guard<std::mutex> lock(mutex_);
   if (started_ && seq < next_seq_) return;  // too late to matter, drop
   buffer_[seq] = std::move(frame);
 }
 
-std::optional<std::vector<int16_t>> JitterBuffer::pop() {
+std::optional<std::vector<uint8_t>> JitterBuffer::pop() {
   std::lock_guard<std::mutex> lock(mutex_);
   if (!started_) {
     if (buffer_.size() < target_depth_) return std::nullopt;
@@ -50,7 +50,7 @@ std::optional<std::vector<int16_t>> JitterBuffer::pop() {
   }
 
   consecutive_misses_ = 0;
-  std::vector<int16_t> frame = std::move(it->second);
+  std::vector<uint8_t> frame = std::move(it->second);
   buffer_.erase(it);
   return frame;
 }
